Move slow test to be integration test (#179419)
Move slow test to be integration test (#149712)
